bug fix - bit test was writting the dists to a wrong directory when dist attribute is set in bit.json
